+++ src/includes/templates/http/html/status/500.php
@@ -2,11 +2,16 @@
 declare (strict_types = 1);
 namespace WebSharks\Core;
 
-/*
- * Merge w/ defaults.
- */
+use WebSharks\Core\Classes;
+use WebSharks\Core\Classes\Core\Base\Exception;
+use WebSharks\Core\Interfaces;
+use WebSharks\Core\Traits;
+#
+use function assert as debug;
+use function get_defined_vars as vars;
+
 $¤defaults = [
-    'Exception' => null,
+    'Throwable' => null,
 
     'http/html/header.php' => [
         'head' => [
@@ -15,18 +20,14 @@ $¤defaults = [
     ],
 ];
 extract($this->setVars($¤defaults, $¤vars));
-/*
- * Output template contents.
- */ ?>
+?>
 <?= $this->get('http/html/header.php') ?>
 
 <h1><?= __('500 Internal Server Error'); ?></h1>
 <p><?= __('This error has been reported to system administrators. Very sorry! :-)'); ?></p>
 
-<?php if ($this->App->Config->©debug['©er_display'] && $Exception) : ?>
-    <pre>
-        <?= $this->c::escHtml($Exception->getMessage()) ?>
-    </pre>
+<?php if ($this->App->Config->©debug['©er_display'] && $Throwable) : ?>
+    <pre><?= $this->c::escHtml($Throwable->getMessage()) ?></pre>
 <?php endif; ?>
 
 <?= $this->get('http/html/footer.php') ?>